pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "email-unsubscribe-agent"
version = "0.1.0"
description = "LLM-powered email unsubscribe automation tool"
authors = [
    {name = "Robert Thompson", email = "rthomp@stanford.edu"}
]
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "beautifulsoup4==4.13.4",
    "cachetools==5.5.2",
    "certifi==2025.7.14",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "coverage==7.9.2",
    "distro==1.9.0",
    "fastapi==0.116.1",
    "google==3.0.0",
    "google-api-core==2.25.1",
    "google-api-python-client==2.176.0",
    "google-auth==2.40.3",
    "google-auth-httplib2==0.2.0",
    "google-auth-oauthlib==1.2.2",
    "googleapis-common-protos==1.70.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httplib2==0.22.0",
    "httpx==0.28.1",
    "idna==3.10",
    "iniconfig==2.1.0",
    "jiter==0.10.0",
    "oauthlib==3.3.1",
    "openai==1.96.1",
    "packaging==25.0",
    "pluggy==1.6.0",
    "proto-plus==1.26.1",
    "protobuf==6.31.1",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "Pygments==2.19.2",
    "pyparsing==3.2.3",
    "pytest==8.4.1",
    "pytest-cov==6.2.1",
    "python-dotenv==1.1.1",
    "requests==2.32.4",
    "requests-oauthlib==2.0.0",
    "rsa==4.9.1",
    "sniffio==1.3.1",
    "soupsieve==2.7",
    "starlette==0.47.1",
    "tqdm==4.67.1",
    "typing-inspection==0.4.1",
    "typing-extensions==4.14.1",
    "uritemplate==4.2.0",
    "urllib3==2.5.0",
    "uvicorn==0.35.0",
]
requires-python = ">=3.10"

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["*"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py", "*_tests.py"]
python_functions = ["test_*"]
addopts = "-v"
